Rework main Tpm structs and traits
Now instead of having a Driver trait and a Tpm struct, we just have the
Tpm trait, which includes the previous Driver functionality, as well as
buffer management.

The run, and run_with_auth methods are now just `dyn Tpm` methods,
allowing them to be _called_ on any object that implements `Tpm`, but
not to be overridden.

This also finishes the implementation for processing command/response
authorizations.

Signed-off-by: Joe Richey <joerichey@google.com>